Talos: Fold stepsState into appState
Those two being separate is a hindrance, as the app currently
maintain state to know which page is displayed using two methods:
isOrderOn (if 'true', we're past the 'preview') and 'currentIndex',
which is an index into the configuration panel.
Instead of a two-level routing we are switching to a single-level routing,
which require folding the steps into the app state so that step 0 can become
the 'preview' step.
