+++ talos/src/components/banManagement/banManagement.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { steps } from "./../steps/static"
 
-import { withStepsState } from "../steps/Container"
+import { withAppState } from "../steps/AppState"
 
 import { isCurrentStep, buildStepClassName } from "../../services/service.step"
 import BanManagementContent from "./banManagementContent"
@@ -78,4 +78,4 @@ class BanManagement extends Component {
   }
 }
 
-export default withStepsState(BanManagement)
+export default withAppState(BanManagement)
