+++ talos/src/reducers/index.js
@@ -3,7 +3,6 @@ import { combineReducers } from 'redux'
 // import { reducer as notifReducer } from 'redux-notifications'
 
 import appStateReducer from './appStateReducer'
-import stepsStateReducer from './stepsStateReducer'
 import secretSeedReducer from './secretSeedReducer'
 import networkOptionsReducer from './networkOptionsReducer'
 import banManagementReducer from './banManagementReducer'
@@ -17,7 +16,6 @@ const reducers = combineReducers({
   banManagement: banManagementReducer,
   administrativeInterface: administrativeInterfaceReducer,
   appState: appStateReducer,
-  stepsState: stepsStateReducer,
 })
 
 export default reducers
